updated generator

+++ src/generator/generator.cpp
@@ -1,8 +1,8 @@
 #include "generator.h"
 
-DrunkEngine::Random::Random(const uint32_t seed)
+DrunkEngine::Random::Random(const uint64_t seed)
     : generator(seed) {}
 
-uint32_t DrunkEngine::Random::NextRandom() const {
+uint64_t DrunkEngine::Random::NextRandom() const {
     return generator();
 }
